[Dsl,.kts] run build notification tests
The build notifications in the Dsl model are not richly implemented or
extensively used yet, but this might change in the medium-term future
as we implement more lint-like checks and quickfixes through the
richer parse of the Dsl model (rather than bare text or
Groovy-specific checks).  Strengthen the current tests somewhat, and
implement the Kotlin equivalents so that we have a baseline of
functionality.

Bug: 138277630
Bug: 142114586
Test: existing tests strengthened and run for KotlinScript
Change-Id: Ic0ad09bc767de256c07fc3d4642dc74f995eb548