Add ComposePreviewSettingsEditor to edit run configurations
Create ComposePreviewSettingsEditor, which represents the UI to create
and edit ComposePreviewRunConfiguration in the run configurations edit
panel. The UI should be fairly simple and contains a module selector and
an edit text where the user can provide the composable fully qualified
name. For configurations created from a specific function, the fields
should be automatically filled accordingly. A debug tab was also added
by using AndroidDebuggerPanel, so we can configure the same debugging
options as Android app configurations.

Bug: 144500928
Test: ComposePreviewSettingsEditorTest
Change-Id: I899241f9fdc7b9f2f3db6f9df6c23a0245911815