Remove getRenderableBoundsForInvisibleComponents
Before this CL, the responsability of finding the boundaries for
SceneViews was in the layout manager. This caused a number of
interdependencies, including having this method in the layout manager
and the surface.

Since now the SceneViews paint themselves, we can move the
responsability of finding the boundaries to only the surface.

Test: Adds new tests for scanlines. Existing tests pass
Bug: 148859613
Change-Id: I48f9b4f4c05f6bd7435982bd2046054c1910bf79