[Dsl] improve tests of GradlePropertyModel
The tests of GradlePropertyModel exercise many basic parts of the Dsl
model: enough that it's worth trying to run them as much as possible
despite variations of the surface syntax between Groovy and
KotlinScript.

The major difference in expression that needs working around in these
parameterized tests is that in Groovy it is possible to define
variables in the ext scope, because that is a model object with a
configuration closure; in KotlinScript, there is no such ext scope:
all extra properties are defined through assignment or delegation.
This means that tests which rely on being able to find ext variables
in Groovy have to be adapted to find top-level variables in
KotlinScript.

Going through the exercise of enabling as many tests as possible
reveals a number of minor bugs in the KotlinScript parser and writer,
now logged in the issue tracker.

Bug: 138277630
Bug: 142114586
Test: updated, strengthened, more run for KotlinScript
Change-Id: I20bf69ae5359bf9df53ab65021c8703b96eefa2e