Update how images are created
When created using a PixelInterleavedSampleModel the image wasn't
hardware accelerated, and so needed to be repainted into a separate
image in ComponentTreeLoader.combine. Now we create an accelerated image
right off in SkiaParser, so the repaint in combine is no longer needed
(unless we're actually combining).
This results in a performance increase of 50-60% for
ComponentTreeLoader.loadComponentTree in my testing (though
we still have a long way to go).

Test: no behavior change, no test change
Bug: none
Change-Id: Ia524f251fa4d6c3fbce62f90f2e4313dbd0e5ee9