Reduce code dup and fix inconsistent heap dump displaying
This commit reduces duplicate code in displaying attributes in heap dump
views (classifier view, class-set view, and instance details view).
It also marks as final properties in `ClassifierSet` that were never
overriden and wouldn't make sense for subclasses to override,
which in turns allows slightly further simplifications.
The refactoring also reveals and fixes a minor inconsistency
in displaying "native", "shallow", and "retained" when it comes
to INVALID_VALUE (-1).

Bug: n/a
Test: fix inconsistent test
Change-Id: Ieed9a98a67006d790252f4505fdc9a2ed1a4dc76