Make icon downloads interruptable
When the IconPickerDialog is closed, it shutdowns the background thread
used to copy and download icons, however, it doesn't guarantee that
running tasks will be interrupted. Potentially leaking any of these
tasks that may take a long time to execute.

This makes it so that the download checks for the state of the current
progress indicator between downloads.

It also changes the download directory for the metadata to the temporary
directory. The metadata in the Android/Sdk directory should always only
represent the current icons present in that directory. So the downloaded
metadata file should be present somewhere else.

Bug: 148819905
Bug: 148559974
Test: updated
Change-Id: I604aa71c6b690c2c56370e714b279a6fb348040b