Add options to trace Gradle sync
Adds three components to Experimental form to enable and
customize trace profile during Gradle sync.
It consists of a checkbox to enable/disable trace, a dropdown
box to select which trace profile to use between default profile
and specified file from disk, a text field with file browser to
select profile.

The dropdown box is enabled only when checkbox is selected, file
browser is enabled when checkbox is selected, and using profile
from local disk.

Bug: 149320690
Test: ExperimentalSettingsConfigurableTest
Change-Id: I9eb9a5b4cc2ee926c2e0b9420acafcb0023c2851