[Dsl] make GradleDslFile.parse() linear
The call to resolveAll() after parse() led to quadratic behaviour,
because it attempted to resolve all unresolved injections through the
entire project at the end of every build file parse, including many
injections which could not possibly have meaningfully changed.

The motivating comment referred to resolving sibling dsl entries, for
example

  ext.vars = [
    key: "value"
    key1: ext.vars.key
  ]

but this fragment is not legal in current Gradle: in isolation this
does not run, because ext.vars does not exist at the time that the
value for the key1 entry is computed.  (If it did exist because of a
previous declaration, it should take on the value from that
declaration rather than "value").

It seems also that Gradle configures and executes builds of parent and
sibling projects independent of the child / sibling build files, so
that attempts to affect the resolution of a parent build file from a
child will not work.

However, with our current one-pass parsing implementation, we must
perform a limited resolution after parsing because on configuration of
the build, a buildscript block's effects happen before the rest of the
build file even if the block is lexically after portions of it.
Restricting the re-resolution to injections whose origin element is
the GradleDslFile in question (and any applied files) means that each
parse pays the cost of its own unresolved variables, and not the
unresolved variables of the entire project.

Bug: 149029879
Bug: 149206270
Bug: 149286032
Test: existing (change in speed, not functionality)
Change-Id: Ic5f7d27733efe3bfa3324bc22833d3149974ba61
(cherry picked from commit 858671b34dfc3ed07a4cc153f740488f433f4321)