[Dsl] show sync banner less for changes in comments
Don't treat a build file as modified and needing a project sync if the
only change is to comments.

This isn't a panacea, because it is possible that the user in making
edits goes through a transitory state where the document is modified
in such a way that it would require sync, and then back, but the
state that a build file is modified enough to require sync is sticky
except when the file hashes coincide.  The more the editor imposes
structured editing (e.g. by inserting comment characters automatically
when the user hits Return in the middle of a comment line) the less
likely the banner is to show up spuriously; if a user inserts a
line comment by hand, the first '/' character will cause the file to
look modified, so the requires sync banner will show.

Bug: 145235353
Test: manual, added to GradleFilesTest
Change-Id: Idaebb14b23aa6fc7bb4fb66c640a15701e807693