Migrate DesignSurface to use Swing layout
This is step 1 of removing some of the duplication we have for layout
and sizing of views in the Design surface. We have a parallel mechanism
to measure the size and to decide the scroll area.
This change centralizes a lot of that into a new SceneViewPanel that,
eventually will handle everything using Swing LayoutManager. For now, it
delegates the layout and positioning of SceneViews to the existing
SurfaceLayoutManager. In a follow-up CL, the SurfaceLayoutManagers will
be converted to regular Swing LayoutManagers.

The CL also simplifies a bit the layers in the design surface and the
layout of the zoom controls.

Bug: 148859613
Test: Existing tests pass and manually checked
Change-Id: Ic510d540ee3b8c7cf7695952b54b2ffc256276f5