Use getResourceModuleDependencies() in BindingScopeEnlarger
For performance reasons, Bazel projects are modeled in IntelliJ like so:

1. All source code belongs to a single "workspace" module

2. There is a module for each resource-generating Android target which
   contains just the resources from that target.

3. To make sure resources are accesible from source code, the workspace
   module depends on every resource module.

4. To make sure sources (like custom views) are accesbile for resources,
   every resource module depends on the workspace module.

The circular dependency created by 3-4 leads to stack overflow when
recursing over module dependencies with
ModuleRootManager#getDependencies, and this causes constant exceptions
when BindingScopeEnlarger is applied to Bazel projects that use data
binding.

This CL resolves the issue by migrating BindingScopeEnlarger to use
the build system-independent
AndroidModuleSystem#getResourceModuleDependencies instead of
ModuleRootManager. For Bazel projects, this means we rely on dependency
info directly from the build system (as opposed to the IntelliJ model).

A subtle difference here is that getResourceModuleDepdencies
already handles getting the transitive closure of module's
dependencies, making the recursion unnecessary. For Gradle projects,
this means we're switching from using ModuleRootManager to
AndroidUtils#getAllAndroidDependencies under the hood. This should
improve performance with caching, but I'm unsure of the behavior
with respect to excluding test modules (though all existing
scope-related tests still pass).

Bug: 135564124
Test: DataBindingScopeTest still passes. Manual testing shows no
 exceptions for Bazel projects.
Change-Id: If1655a854fb5e0fa9d9c0f1fad8846cc5cc5fa99