[Sync] Give sync state banner a dismissal option
The dismissal hides the current notification banner.  A related
banner (e.g. the "You can edit in the PSD" message) may reappear after
a few keystrokes if the user has not configured that other banner to
be hidden; because the dismissal does not update the file hashes (as
this might lead to loading an incorrect cached sync state) subsequent
minor edits can lead to the banner showing up again in a perhaps
surprising fashion.

Bug: 145235353
Test: manual

Change-Id: I9b6f005ff7479a32f480e90fb52c9cc3ec02e4aa