Fix primary key bug in JdbcDatabaseConnection.
JdbcDatabaseConnection uses PRAGMA table_info(tab) to get primary key information.
If a column is part of the primary key, that command returns an integer for that column
that corresponds to its position in the primary key.
So if a table has two columns in the primary key, the first column will be 1 and
the second 2.

Therefore to know if a column is in the primay key we should check if the number
return from table_info is bigger than 0.

Test: JdbcDatabaseConnectionTest.kt
Bug: 149390817
Change-Id: I3c63b87b41698b41af4aa8884f3926bbc574e83f