[PerfettoCapture] Refactor CpuCapture.
This CL refactors CpuCapture to include all the data-related methods
from ATraceCpuCapture into it, so we can then use CpuCapture as a base
to PerfettoCapture (to decouple ATrace with Perfetto) and unify most of
our code base to pass CpuCapture where we can.

Also changes all the "instanceof" checks with more getType() ones.
When PerfettoCapture lands, we will probably move to a more sophisticate
API to check if some particular data/feature is available for that
capture object.

This breaks the CpuCapture into the public interface and the base
implementations so it's more clear what is supposed to be override and
to make it easier to look at the API without mixing impl details.

Bug: 147099951
Test: existing
Change-Id: Ic070cacf961b36a46156fefdbf57c943a7aa835d