[Gradle Sync] Remove ProjectJdkSetupStep
I have no idea what this is actually doing, we already verify the Jdk
before sync and sync should not have any effect on the project level
JDK. We set up all Android modules to have their own SDK. Java modules
use the project based one, but as before this is verified before sync.

It is also worth noting that setup step may seems to be able to
(although I'm not sure of the exact situation this would happen) change
the Jdk to one that the user didn't select. This is due to using the Jdk
attached to the first Android Sdk it finds, in most cases this is the
correct one but it is possible for some of the Sdks to have the wrong
Jdk. Whether it is possible that this one can be picked is still up in
the air.

Bug: 149772019
Test: N/A
Change-Id: Ic971dd6d61ab72b395c63efbd003a8092c1183a1