Refactor descriptors
The goal of this CL is to take away user's ability to create descriptors that allow them to attach to targets. Descriptors should only be provided by discovery service. We also need to inject certain dependencies (namely AppInspectionJarCoper) so that everything we need to launch a target is inside it.

In order to do so, I refactored our descriptors by extracting an abstract class called ProcessDescriptor. It contains overriden equals and hashCode function so that descriptors from different sources can be compared and matched. All concrete implementation descriptors now inherit from ProcessDescriptor.

Test: updated tests
Bug: 149308030
Change-Id: Icb24d9e124c1d5fb4978e1f288211cfc184fcf4b