Rename misleading SourceProviders.allSourceProviders
to `currentAndSomeFrequentlyUsedInactiveSourceProviders` with the
further intention of deprecating them when it becomes possible. The long
descriptive name should for now ensure that the collection is not used
accidentally instead of `currentSourceProviders` collection.

It is currently infeasible to support `allSourceProviders` in Gradle
project system as the AGP does not return all for performance
reasons. No other project system support build variants and the
collection matches `currentSourceProviders` in non-Gradle project
systems.

Bug: n/a
Test: n/a
Change-Id: I9b51c8d789b64802b16e6898cf18dbff7622c934