[help panel] Add listener to the navigation button
The tutorial we made is all navigatable through the navigation
button, and it fits on a single page from the start.

So we can't really use the "scrolled to the bottom" to understand
the user involvement.

I figured maybe better user metrics would be to understand how often
user navigates through the tutorial. This CL adds that listener
component.

Let us know what you think :)

Bug: N/A
Test: Manually tested
Change-Id: I1dd99ae165dec4a0a0c2911ef69da73a327a1fa0