[Gradle Sync] Remove module links jumping to element
This change removes the ProjectBuildModel being passed to
SimpleDeduplicatingSyncIssueReporter#createModuleLink, and also removes
any functionality that it was being used for. This changes all module
links to no longer jump straight to the element, while this is nice to
have it is not very impactful and not worth the performance penalty.

If we later decide that we would like this then it should be added back
to the module link handlers rather than.

Bug: N/A
Test: adjusted
Change-Id: I44af2246b25e598c1d46fc8cd322f28f47aba41d