Create a Swing peer for SceneView
This allows for the SceneView to have Swing components. This CL also
moves the ModelNameLayer to be a Swing component to prove the
functionality.
This will also allows to support for arbitrary Swing components to be
used as part of the SceneView painting. As part of the CL, a mock
toolbar has been added behind a flag to test the toolbar functionality.

Bug: 148859613
Test: Manually tested
Change-Id: Iefd543421db70b11bd4ba02700a3ac1ca7f254d5