[Dsl] begin handling global imports
Gradle imports a number of things into the toplevel of build files; a
list is available in the Gradle manual under "Default imports".  The
current Dsl parser supports only org.gradle.api.rootDir and
org.gradle.api.projectDir; in preparation for supporting more of these
imported properties to reduce the occurrence of unresolved
properties (with all their scaling implications), change the
representation from code at resolution time to data in the
GradleDslFile.

This requires overriding some lookup functions in
GradlePropertiesDslElement; most of the changes in this commit are
changes to access modifiers and rearrangement to provide only two
functions to override.

Bug: 149029879
Test: existing (no change in functionality)
Change-Id: Ib5440455757c6720e84dd8643ce069a53ebab8a9