[Dsl] test expansions and improvements
Continue porting tests written for one build language to the other,
where appropriate.  (In some cases this means writing a Groovy test
for an origially-Kotlin test case).

Annotate some tests with assumeTrue() and a human-readable reason for
not running that test for a particular language, rather than using the
implicit assume in writeToBuildFile().

Bug: 138277630
Bug: 142114586
Test: included
Change-Id: Ifb57164b04aa80db072bc48c6d14ba8e302bce87