Disable the native row selection
JTables have native row selection. Checkboxes are more user friendly. The
old Modify Device Set dialog checkboxes deferred to the native row
selection. But the behavior of both selection mechanisms was different.
Clicking a checkbox is enough to add the device to the selected set but you
needed to Ctrl+click to do the same with the native selection.

Remove the inconsistency by disabling the native selection and only using
the checkboxes

Bug: 148897210
Test: ModifyDeviceSetDialogTableTest and ModifyDeviceSetDialogTest
Change-Id: I656bbebfcbc2887e63ef5fa9257112b4689203ab