Unlock emulator features in Crostini w/ /dev/kvm
With the introduction of nested hardware assisted virtualization
(/dev/kvm) in Crostini (starting w/ Chrome OS M82), machines that
support this now behave more like Linux machines wrt Studio
emulator functionality.

On machines running Crostini that support /dev/kvm, Studio now
behaves much like Linux (emulator is installed, and AVD manager
functionality is fully available). However, on machines running
Crostini that don't support /dev/kvm, Studio reverts back to the
old behavior on Chrome OS, which is that the emulator (in Android
SDK tools) is not installed, and AVD management functionality is
not made available.

This change also fixes a bug on Chrome OS devices that do not
support /dev/kvm, which would get a "SDK emulator directory
is missing" error message when Studio was run for the very first
time and the Android SDK needed to be installed. Basically, the
error results from the fact that the emulator tools wouldn't be
installed, in the first place, but then a check for the `mksdcard`
binary was made.

Fixes: 145151771
Test: QA team manual testing
Change-Id: I35951e5260810e53d48ca5ca25a800efc2b6efd8