Update sync tests for Gradle 6.2
Update composite build sync output. Also, there
is an issue in Gradle 6.2 which means that groovy-all
is not in the compile classpath for buildSrc. This does
not affect autocompletion in .groovy files, but it does
not resolve symbols in .java files. Because this is not
a frequent thing to have, we are going ahead with this
change. Issue is reported to Gradle
https://github.com/gradle/gradle/issues/12274.

Bug: n/a
Test: existing
Change-Id: I0cdbb1f1ea17737da5526cc126982edcb8a2896b