Return module from LintResult instead of hardcoded `null`
This CL re-adds the behavior that was missed when BatchLintClient and
EditorLintClient was reconciled to use the base LintIdeClient.
LintIdeClient returned default `null` in `LintIdeClient#getModule`
method. This behavior was overriden by EditorLintClient to return the
correct module (See changes to `AndroidLintIdeClient` in ag/I7351ab918b).

This CL delegates `getModule()` to `LintResult#getModule()`. Only
`EditorLintResult` overrides this method, so there should be no
side-effects.

Test: Added `testLintIdeClientReturnsModuleFromEditorResult` to ensure
that correct module is returned to `LintIdeClient` when
`LintEditorResult` is used.
Fixes: 149912107

Change-Id: I1914a7a82a137e8d04c0e4a7493165a68f622340