Remove Preview Tool Window in favor of Split Editor
The Split Editor was introduced in Android Studio 3.6, making the Preview
Tool Window obsolete. Now that we're starting to be in Android Studio 4.1,
we can remove the split editor flag, enabling it permanently, and remove
all the code associated with the Preview Tool Window.

This change also removes the NELE_SPLIT_EDITOR flag since it's not
needed anymore.

Bug: 63923598
Bug: 146399022
Bug: 149990005
Test: Unit + UI tests
Change-Id: I1a9acfe1b117a93089621c8811fc4b78e68d1f4b