Replace NlModel.create methods with a builder
This replaces all the different .create methods with a builder. A
follow-up CL will replace the SyncNlModel test builder with this same
one, unifying all the model building in one place.
The only NlModel#create method left in the interface was left there as
it is being used in the NlPreviewForm that is currently being removed.
Left there to avoid the conflict.

Bug: N/A
Test: Existing tests pass
Change-Id: Iedabc5ffd3dc314e2473e5032398a8023d238764