Trace IDE part of Gradle sync
To trace Gradle sync for IDE, Studio should be launched with
javaagent argument.
When trace option is enabled or disabled, show dialog to ask
users for restart. Since intellij doesn't provide interface
to inject jvm args on next restart, we add/remove the argument to
vmoptions file, so that the next launch of studio will have
expected arguments.

Also update the way to inject jvm args to Gradle, it now reads
the launch arguments of Studio, and trace Gradle only if Studio
was launched with javaagent.

Bug: 149320690
Test: included.
Change-Id: Ib8aff44d4d71c47286d0e5323bb431167837110d