[templates] apply changes to build model only once
The new DefaultRecipeExecutor2 executor applies changes to the build
model at the end of all modifications; this is appropriate because it
has a single build model.  The older DefaultRecipeExecutor executor
applies changes after each distinct modification, which is acceptable
because it generates a new build model each time.

The transitional RecipeIO structure attempts to broker this difference
in behaviour, as well as enabling dry runs, but its applyChanges()
method unconditionally applies changes to the build model, which under
some conditions can invalidate Psi elements, probably racily (given that
this happens when Studio runs normally, but not under debug).  Make
sure that the RecipeIO object for DefaultRecipeExecutor applies changes,
and the one for DefaultRecipeExecutor2 does not.

Fixes: 150113098
Test: existing, manual
Change-Id: I55d98a46ab7f0cfdc2c171b0536c75ea9e39c3f6