Fix Binding classes not resolving due to dumb mode
Earlier, in an attempt to guard against an exception thrown by the data
binding / view binding cache system, we may have papered over the
original bug in the wrong way, by blocking the exception from being
thrown at a low level in our system.

This introduced a new problem: the IDE would ask us to update our
binding cache, but due to being in dumb mode, we would get back
incomplete intermediate results, which would then be saved into the
cache.

Then, later, when callers tried to query the latest binding information,
even when they weren't in dumb mode anymore, they would get stale
results.

The main before / after difference is logic that went from this:

  if (lastModificationCount != resources.modificationCount) {
    lastModificationcount = resources.modificationCount
    ... update cache ...
  }

to this, where we only update the modificationCount after we are sure
the update went through:

  if (lastModificationCount != resources.modificationCount) {
    ... update cache ...
    lastModificationcount = resources.modificationCount
  }

We also entirely remove the IndexNotReadyException guard in
BindingXmlIndex, as it should no longer be necessary. All of the
original callstacks pointed to the cache system that we fixed here.
If there are any other leaky areas, I actually want to get more
exception reports to highlight them.

Fixes: 149630915
Bug: 147910426
Test: Added
Change-Id: Iab0b3f3b4b9c71d5b45a2dbf7f3393c68e255b8b