Clear cache when "viewBinding { enabled }" setting is toggled
We currently cache all the modules with view binding enabled in them,
but that list was not getting refreshed when a new module had its
viewBinding setting toggled.

Here, we copy the pattern used by data binding, which already handles
this case, and make a version for view binding.

Note: We also take advantage of this CL to remove some tangentially
related but obsolete TODOs.

Bug: 119823849
Fixes: 150405777
Test: Added
Change-Id: I6d033e0dfa68f6e83260058c806498948363edbe