[Windows] Remove non-execute prefix in test JSON
Gson does not properly handle the JSON non-execute prefix with Windows
line endings, so IconPickerDialogTest fails when icons_metadata_test.txt
is checked out with Windows line endings because Gson thinks the JSON is
invalid. This won't happen in production because the icons_metadata.txt
file packaged into the JAR has Unix line endings, so this CL just
removes the prefix in the test file as a workaround.

Bug: 147453294
Test: IconPickerDialogTest
Change-Id: Id22883b6c76cd8fcea5aeb5da864a6b4e2b9dc53