[Gradle Sync] Don't request artifacts from Gradle more than once
This was causing issues in IDE setup as we requested the same extra
artifacts (sources/JavaDoc) multiple times. This resulted in duplicate
entries in the returns list. The IDE setup code was written with the
assumtion that there would never be duplicates in this list.

This change ensures we de-duplicate before requesting. First it
removes an anonymous object that doesn't work well with comparisons, and
then adds a call to distint to the requested ArtifactIdentifiers.

Bug: 150402775
Test: manually tested
Change-Id: I488d43b2f38123a18f3f1de197b4d60ad598d611