Add and update rows through diff.
Doing a diff has the advantage that we can update
individual cells without recreating the UI.
The issue with recreating the UI is that each
time the user edits a value in the table
the row that is currently focused is removed, therefore
after the edit the user has to re-navigate there.
Which is not a great UX, in particular
when using the keyboard to navigate the table.

Test: TableControllerTest.kt, TableViewImplTest.kt
Bug: 150628254
Change-Id: Icfbff7d4db8d28ce0570a68a82c1ad26fa1afd54