Fix stale cache issues when indexNotReadyException
Do not cache data if it's dumb mode.

Issues happen when we query index in dumb mode, since it's already
in read action, we return null. And this is cached by accidient.
In this fix, we return previsouly cached data if indicies are not
available. Fresh data is provided and cached in the next call, once
it's smart. And this is guaranteed because high frequency of invocations.

Bug: 150490926
Test: DumbModeTest
Change-Id: I0db4548756cf8d278ffdbea7a0efa3171e0b60c9