Fix a bug that made tables without primary key not editable.
If a table has no primary key, we rely on rowId to run
updates.
In a recent change we went from storing SqliteRows
in the model of the table, to storing just Strings.
That CL made an attempt to recreate the SqliteRow
before passing it to udateCellInvoked.
This introduced the bug, since we don't display
rowId in the UI the row was not recreated correcly.

This was fixed by passing the index of the row
instead of the row itself.

Test: TableViewImplTest.kt, TableControllerTest.kt
Bug: 150756511
Change-Id: If122cbc47c7166f4955f06dce20ab7c3fb5275a2