[Dsl,Ps] rename signing config references
Although this might look like a sound solution, and this does seem to
function as intended, there is a wrinkle.  It is possible and
reasonable to refer to signingConfigs, such as

  android {
    signingConfigs {
      z {
        storePassword 'ab'
      }
    }
    buildTypes {
      release {
        signingConfig signingConfigs.z
      }
    }
  }

and this is in fact the form we recommend at
e.g. <https://developer.android.com/studio/build/build-variants>.  Not
(absolutely not!) recommended, but legal, would be something like

  android {
    signingConfigs {
      z {
        storePassword 'ab'
      }
    }
    buildTypes {
      release {
        signingConfig signingConfigs.z
        applicationIdSuffix ".${signingConfigs.z.storePassword}"
      }
    }
  }

and this rename implementation currently fails to cater for the use of
the signingConfig property as an rvalue.

Bug: 147867235
Test: added a PsSigningConfig test
Change-Id: I4255b76b773f2f926c3a4807aa271bc0cc23dc39