Add tests for separate heap-dump UI
This CL clones the old tests for legacy heap-dump UI and adjusts
it for the new UI. It exposes more high-level states in the new UI
(e.g. is showing capture object, etc.) to enable equivalent assertions
from the legacy one without being too specific (e.g. second component
of the splitter being null).
It also finds and fixes a null-pointer error when loading
the heap-dump from a file.

Bug: n/a
Test: added new
Change-Id: Ic1ce869223c4323cda05aa2296adcc5de03bd166