Don't hyperlink single slashes to root
The Bazel plugin uses slashes to indicate sync progress in an output
console (e.g. "[123 / 4,567]"). This change stops GenericFileFilter
from creating a hyperlink to the file system root for these slashes.

The downside is that this also stops the console from hyperlinking
legitimate references to the root directory, but we suspect this case
is much less common than forward slashes which shouldn't be
hyperlinked.

Fixes: 150688172
Test: updated GenericFileFilterTest
Change-Id: I16e875bc122c2f04cae6b76e6c2c48f2991412f3