[BA] Generate one report for all task issues
Change report generated on 'generate report' link click to be
generated from the task for all issues with this task at once.
It is expected to be a rare case when there are more than
one issue for the task. This change is required for inlining
task issues to one task page and having only one
'generate report' link there.

Bug: 150858372
Test: Updated
Change-Id: Icdcb77517b6a5c9b128ea0bb1f21cfc98cdec478