Add a concept of TableSupport in the property panel
When a property editor that is used as a table cell editor there are
2 events where the table must be notified:
 - the editor is taller than a standard table row
 - some navigation events are intercepted by the editor component

Until this CL a custom solution was used to implements these
features. With the CL these concepts are moved into a TableSupport
instance that are created by the cell editor provider and supplied
to the property editor model. This is a better approach that can be
expanded to other events in the future.

Remove the old support including PTableVariableHeightCellEditor.

Bug: 150322359
Test: Unit tests are updated.
Change-Id: Ia545b2a29d2748e75824b7f56bd67fb17ef1c1d6