Refactor and expose filter's test on each InstanceObject
This CL makes implementing an instance filter a matter of
passing in parameters instead of overriding methods.
It also exposes the predicate for testing whether some individual
`InstanceObject` passes the filter (for use in a future CL),
and factors out the duplicate code that runs through the
instance list to filter the instances.

Bug: 149595994
Test: existing
Change-Id: I25b9772f091faaec53222cfccce6dcaa01d4ed42