Fix UI tests timing out on Mac
Some UI tests are timing out locally on Mac waiting for sync event. These
tests rely on IdeFrameFixture#waitForGradleProjectSyncToFinish, which in
turn calls UiTestUtilsKt.waitForIdle. This method heuristically waits for
idle by posting a key press and release events and waiting for them to
show up in IdeEventQueue.

The problem on Mac (FWIW, with touchbar) was the key chosen: VK_PAUSE.
Since macbooks do not have this key, VK_PAUSE is not properly mapped and
we wait indefinitely for the events to show up in IdeEventQueue. Changing
the key to a common one (VK_F1) seems to fix the issue on Mac while keep
the tests passing on Linux.

Bug: N/A
Test: UI tests
Change-Id: Ifdefd972bf001163d58e8663833ecdc5a3c48d42