[Dsl,.kts] improve moveDslElement
Under some circumstances, adding a newline to a Kotlin file using the
Psi interfaces does not actually add a newline: the Kotlin
autoformatter removes it again.  The immediate symptom of that is that
the added PsiElement is invalid, or (as here) if that's not tested
for, subsequent wrong computations including deleting the entire file.
A string is more likely to survive file normalization and is unlikely
to alter the syntax of surrounding elements (unlike, for example, a
line comment).

This shows up particularly in KotlinScript Dsl tests of extra
properties, because unlike in Groovy build scripts, there is no ext
Dsl block as such, so determining a local anchor is not completely
straightforward.

Bug: 138277630
Bug: 142114586
Test: add PropertyOrder.renameReordersProperties test
Change-Id: Icb00d6f0e836720f8d02ba8c73dcec219f5f1b99