Refactor AppInspection's process discovery
We now obtain preferred processes in a one-time operation. They are
basically the applicationIds of the modules in the project. Because of
this change, AppInspectionLaunchTaskContributor can be removed along
with the code it calls into in discovery. AppInspectionDiscovery is now
much simpler.

As a result, discovery service can now discover processes that are
manually launched on a device, without the user pressing "play", "debug,
etc.

In addition, the criteria that deterimine whether a process is
inspectable has changed. Now, the process must be 1) debuggable and 2)
running on an O+ device. (process is guaranteed to be debuggable if we
get it from transport pipeline).

Changed AppInspectionHostService from an application service to a
project service. This makes sure process discovery happens within the
the project and fixes the issue where processes belonging to other
projects are discovered.

Test: New test checks whether pre-O devices are ignored. And test for
prioritization of preferred processes in combobox.

Fixes: 150765564
Fixes: 147777408

Change-Id: I948e0f257fe169129dc95ffa02f4fea3b0a45086