Do not update representations when disposed
Structure window creates a temporary editor just
to call getStructureViewBuilder. Creating an editor
for Kotlin files and some callbacks that can be
executed later and in this case they actually get
consistently executed after the editor is disposed.
We should check for the editor to be not disposed
before executing callback code.

Bug: N/A
Test: N/A
Change-Id: I7c652be07c8f87af07af5138f75d935b63aa2ffc