Mark PropertyTextFieldWithLeftButton as compact.
PropertyTextFieldWithLeftButton is using DarculaTextBorder as its
border. The height of this border varies with 1 pixel depending on
the whether the control is in a table or not.

When we use this control in a variable height editor we often ask
for the height of the cell editor before the parent is set to the
table causing the border to compute the wrong height.

By setting the control to compact it will stay the same height and
we avoid possible flickering.
The flickering is seen after the CL: ag/10554726 which removed a
hack to avoid this in the table implementation.

Fixes: 151148752
Test: None
Change-Id: I63283f4eb30af842cc171a83fb1a8a00b7ed0dec
(cherry picked from commit d82b1c22b04594d59bfcfd1aecd2a361c254b038)