Only set custom row height if there is a change in height
It turns out that when JTable.setRowHeight(row, height) is called
a repaint is required even when the new height is identical to the
current. This causes the properties panel to constantly paint thus
using a lot of CPU for no good reason.
Fix this by only calling setRowHeight when there is a change.

This CL is needed if the refactoring in ag/10554726 is not used.
With that CL we will not need this fix since this code was removed.

Fixes: 151152808
Test: None since this code will be obsolete very soon.
Change-Id: I9c72273f06d47639de408986a2a87b8974888f27