Add support for collections in ParametersBindingDialog.
SQLite can have statements like SELECT * FROM tab WHERE id IN (?),
where the positional parameter has to be replaced with a
list of values.

Unfortunately we cannot rely on prepared statements to
correctly bind arrays, since JDBC does not support it
for SQLite.
Therefore the solution proposed in this CL is to manually
replace the positional parameter with multiple positional
parameters, one for each element in the list, so we can
just bind individual values using prepared statements.

This CL adds the logic for doing that, the UI will be
added in the following CL.

Test: ParametersBindingControllerTest.kt
Bug: 151315656
Change-Id: Ie664b7ec3c5d1658dad1d3c0a2dadc250d366999