[Dsl] new ModelPropertyDescription class
To hold information about model properties.  At the moment this merely
holds a string name, and is unused.  Although it looks a bit like a
Kotlin data class, eventually instances should probably be singletons.

This is motivated by an upcoming change making matchingFallbacks a
mutableList property, which will mean that we can't assign to it using
listOf(...).  We currently hold limited information about types of
things in the Dsl model, but adding yet another piece of information
in dynamic code rather than tabulating it does not feel right.

Bug: 141842964
Bug: 148657110
Test: existing (no change in functionality)
Change-Id: I329b59b712eb9ddbe6d6adb6b732bade97f29699