[Dsl] use ModelPropertyDescription in maps
At the moment this only encapsulates a string: preserve the existing
mechanisms everywhere, and merely unencapsulate where necessary.

This change is noisy in that it mostly replaces one type by another in
a large number of different classes.  It leaves some lines longer than
the standard, which will be addressed in an immediately following
change.

Bug: 141842964
Bug: 148657110
Test: existing (no change in functionality)
Change-Id: I8dd6461f25247b56b3b245a42b72c2bf918b914c