Ensure getDebugBridge() never blocks the caller
getDebugBridge() returns a Future, which is a strong hint that the
method should never block the caller. This CL ensures this is always
the case, even if getDebugBridge has to terminate ADB first. Prior
to this change, terminating ADB would be done synchronously, sometimes
causing UI freezes.

Also re-enable the test on Windows, as there is no reason it should not
work on that platform.

Bug: 150232901, 150163825
Test: Basic tests (we are not really hitting the new code path though)
Change-Id: Ic7001dbf7a35eebfe5419345a57a2c6e1a28793a
(cherry picked from commit dcac6657166b7ca9f2c180fc634b965dea7eee7b)