[Dsl] convert external names to model descriptions
I am fairly convinced that most users of externalToModelName use it
without hierarchy, property by property.  That's the best way to be
correct, I think, and certainly the limited support for hierarchy in
the Groovy implementation worked mostly only by accident: it would go
wrong in principle if someone handed it a named object with dots in it
whose parts happened to look like model objects.

Motivated by the need to annotate Dsl objects with the model
properties that they affect, rather than merely the name of the
property, rename `externalToModelName` to `externalToModelProperty`,
and have it return a ModelPropertyDescription.  Adapt callees to the
new behaviour, at the moment trivially to extract the name from the
property.

Bug: 141842964
Bug: 148657110
Test: existing
Change-Id: Id1eaf9bb265400c4009e18f492a6c169f8358868