Android: fix deadlock during unit tests execution
This is defect in prod code, not in test suite.

Deadlock happens in the following situation: MergedManifestManager is a module service. It contains MergedManifestSupplier as a field (MergedManifestManager#supplier), and clients are using @JvmStatic getFreshSnapshot, which invokes supplier.get().get(). supplier.get() produces a Future (SettableFuture from guava) which is never completed in some circumstances. The cause is as follows: supplier delegates all the work to ThrottlingAsyncSupplier in runtime. ThrottlingAsyncSupplier uses platform's Alarm class (field ThrottlingAsyncSupplier#alarm) to schedule computations. Computation scheduled by the alarm sets the Future, and this releases all the blocked clients. Alarm itself is configured such that ThrottlingAsyncSupplier is its parent disposable. Parent disposable of ThrottlingAsyncSupplier is AndroidFacade. When AndroidFacade is removed from the module, it disposes, and causes disposal of the Alarm. Alarm cancels all the scheduled computations on disposal, and SettableFuture is never set with any result.

In unit tests this situation (removing and adding AndroidFacade) is more common, because tests often share the same module in tests. In particular, c.a.t.i.u.r.ResourceExplorerToolWindowTest followed by c.a.t.i.u.r.r.DrawableIconProviderTest sometimes results in bad timing and deadlock.

Fix is as follows:
1) ThrottlingAsyncSupplier cancels Future on disposal. This releases all the waiting clients.
2) MergedManifestManager is not a module service anymore: now it is a user data in AndroidFacade. MergedManifestManager needs to be re-created each time a module assigned with a new AndroidFacade, otherwise MergedManifestManager will contain stale reference to disposed supplier, which contains a reference to disposed Alarm (which throws exceptions on schedule attempt).

GitOrigin-RevId: 179a7185cb3c418c6cfeedde470f2205fdef1af3