[Dsl] don't unnecessarily test property names
When we have a specific property model (for a specific model
property), it doesn't really make sense to test for the name of the
property: we must have known it in order to get the correct value.
Furthermore, it's more awkward to test for the name of the property as
we further decouple external representation of operators on properties
from internal representation of properties themselves.

Bug: 141842964
Bug: 148657110
Test: updated
Change-Id: I3ad6045fa2505e0121cbe1b3ed1641ba1ca858b6