Only consider connected devices
Execution targets are compared by ID. Note that the active target is not
updated in DeviceAndSnapshotComboBoxAction::updateExecutionTargetManager
when the old and new targets are equal.

Say you had an execution target for a device and the device was later
connected. The new target would never replace the old one because the IDs
were equal.

Change DeviceAndSnapshotComboBoxExecutionTarget to only consider connected
devices

Bug: 150812772

Test: DeviceAndSnapshotComboBoxActionTest
    ::updateConnectedDeviceActiveTargetIsSet

Change-Id: I445a39cfc6018bee82b12cc7ca7cdb2d4b435a3e