[Dsl] centralise the maybeTrimForParent problem
Well, there are multiple problems, but one of them is that the model
hierarchy is not the same as the lexical scope.  At present, we write
out an element name trimmed given its position in the hierarchy, which
is OK but only works if all hierarchy elements have configuration
closure/actions attached.  If any are expressed using dotted notation,
then the trimming is possibly invalid.

Bug: 151607418
Test: existing
Change-Id: I82808020cfaab07016bd9218c616f8047a0336ab